Tune OD_CFL_SCALING4 with a robust regression
Collected blocks in subset3 where OD_CFL_SCALING4
would be applied and that were correlated enough
that NO_REF would not be considered.
Applied the Theil-Sen estimator to each pair of
coefficients to produce a new set of scalings.

BD rate subset1 1.0-5.0 bpp:

            RATE (%) DSNR (dB)
     PSNR -0.0048667 0.000511038
  PSNRHVS -0.00488944 0.000416557
     SSIM -0.00461787 0.000436625
 FASTSSIM -0.0143137 0.000613985
CIEDE2000 -0.0165684 0.00129192

For the Cb plane:

            RATE (%) DSNR (dB)
     PSNR -0.0220416 0.00110778
  PSNRHVS -0.0130546 0.000809958
     SSIM -0.047171 0.00221001
 FASTSSIM -0.00017643 -0.000343799

For the Cr plane:

            RATE (%) DSNR (dB)
     PSNR 0.00132778 3.55187E-05
  PSNRHVS 0.00324951 -7.68438E-05
     SSIM -0.022304 0.00129872
 FASTSSIM 0.382165 -0.0114586

Signed-off-by: David Michael Barr <b@rr-dav.id.au>